+++ src/rotta_rs_module/arrayy/function/ln.rs
@@ -4,7 +4,7 @@ pub fn ln_arr(x: &Arrayy) -> Arrayy {
     let vector = x.value
         .iter()
         .map(|v| v.ln())
-        .collect::<Vec<f64>>();
+        .collect::<Vec<f32>>();
 
     Arrayy::from_vector(x.shape.clone(), vector)
 }
