+++ src/rotta_rs_module/arrayy/function/mean.rs
@@ -1,6 +1,6 @@
 use crate::arrayy::Arrayy;
 
 pub fn mean_arr(arr: &Arrayy) -> Arrayy {
-    let mean = arr.value.iter().sum::<f64>() / (arr.value.len() as f64);
+    let mean = arr.value.iter().sum::<f32>() / (arr.value.len() as f32);
     Arrayy::new([mean])
 }
