+++ src/rotta_rs_module/loss/mse.rs
@@ -9,6 +9,6 @@ impl MSE {
     }
 
     pub fn forward(&self, prediction: &Tensor, label: &Tensor) -> Tensor {
-        (1.0 / (prediction.len() as f64)) * &(label - prediction).powi(2).sum()
+        (1.0 / (prediction.len() as f32)) * &(label - prediction).powi(2).sum()
     }
 }
