+++ src/rotta_rs_module/module/initialization/random.rs
@@ -3,7 +3,7 @@ use rand::Rng;
 use crate::rotta_rs_module::Module;
 
 impl Module {
-    pub fn random_initialization(&mut self) -> f64 {
+    pub fn random_initialization(&mut self) -> f32 {
         let random = self.rng.random();
         random
     }
